Splitting test cases
